% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npm_install.R
\name{boltr_npm_install}
\alias{boltr_npm_install}
\title{Install boltr npm dependencies}
\usage{
boltr_npm_install(package = "boltr", force = FALSE)
}
\arguments{
\item{package}{package name}

\item{force}{Prompt the user to accept install?}
}
\value{

}
\description{
Install boltr npm dependencies
}
